Split up the SQL subscription/projection project
